Added Egg Hatch command
Format:
@Pokemon.Hatch(PartyIndex,[CanRename],[Message])

Description:
If the Pokémon specified in the PartyIndex argument is an egg, it hatches immediately

Arguments:
PartyIndex (int) - Index of a Pokémon in the player's party. Values can range from 0-5.
CanRename (bool) - Optional. Whether the player can rename the Pokémon or not after it hatches.
Message (str) - Optional. The message to display after the Pokémon hatches.