Added @Pokemon.LevelUp(LevelAmount)
This command levels a Pokémon up by the amount of levels (integer) specified in LevelAmount.
If a Pokémon could learn one or more moves on any of the levels between its original level and its level after the command, the command will allow the player to teach each of those moves.