Fix Walk NPC Movement...
Now an NPC with the Walk MovementType will no longer face a direction that is not in its MoveRectangles. When blocked in all directions, it will turn towards a place where it would otherwise have moved towards. Also slightly changed the chance for the NPC to walk or change direction.