Added @System.ReplaceTextures(File) ...
With this command you can add texture replacements similar to how a content pack replaces textures.
This is especially useful when, for example, you have multiple supported languages and want to replace certain textures with localized versions.
The file that is required is in the same format as the exceptions.dat file in content packs, but only supports texture replacement entries.
The replaced texture list is cleared when quitting the current save game.
Content Packs should still work, even with the replaced textures.